feat: Image를 불러오고 저장하는 로직 작성
