+++ server/src/auth/auth.controller.ts
@@ -3,7 +3,7 @@ import { AuthService } from './auth.service';
 import { AuthCredentialDto } from './dto/auth.credential.dto';
 import { RegisterDto } from './dto/auth.resgister.dto';
 import { ApiBearerAuth, ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger';
-import { saveImage } from 'src/util/story.util.saveImage';
+import { saveImage } from 'src/util/util.generate.randomFileName';
 import { FileInterceptor } from '@nestjs/platform-express';
 import { CustomExceptionFilter } from 'src/exception/custom-exception.filter';
