+++ server/src/image/image.controller.ts
@@ -1,15 +1,15 @@
 import { Controller, Get, Param } from '@nestjs/common';
 import { ApiOperation, ApiResponse } from '@nestjs/swagger';
+import { ImageService } from './image.service';
 
 @Controller('image')
 export class ImageController {
-  constructor() {}
+  constructor(private imageService: ImageService) {}
 
-  @Get('profile/:path')
-  @ApiOperation({ summary: 'Get Profile Image' })
-  @ApiResponse({ status: 200, description: 'profileImage' })
-  async getProfileImage(@Param('path') path: string) {
-    return
+  @Get(':path')
+  @ApiOperation({ summary: 'Get Image' })
+  @ApiResponse({ status: 200, description: 'Profile Image or Story Image' })
+  async profileImage(@Param('path') path: string): Promise<Buffer> {
+    return await this.imageService.readImage(path);
   }
-
 }
